p.71
This section contains material from the Ó Cléirigh Book of Genealogies
(OCBG). On one level, this is very interesting, as it will allow us to
compare different versions of the same pedigree. However, I am not
actually sure whether Mac Firbisigh himself added it or whether it was
added by O'Donovan (if it is O'Donovan's addition, presumably we omit
it?). He describes the material thus: "This line of Mac Giolla
Cheallaigh's pedigree is inserted from the genealogical MS. in the
handwriting of Peregrine O'Clery, now preserved in the Library of the
Royal Irish Academy; it comes down seven generations later than the line
given by Mac Firbis" (can I respectfully point out that this kind of
confusion becomes inevitable when scholars feel unable to use the
first-person in academic writing? :-P).

The prose paragraph at the foot of p.71 was definitely added by
O'Donovan from OCBG: "All this matter enclosed in brackets, down to the
end of the pedigree of the O'Clerys, has been inserted from Peregrine
O'Clery's genealogical MS now deposited in the Library of the Royal
Irish Academy. Mac Firbis has omitted this family altogether, but, as it
appears from the authentic Irish Annals that they had supplied many
distinguished chiefs to the territory of Hy-Fiachrach Aidhne, the
Editor, deeming it a pity that they should not have their place among
the families of the race of Guaire Aidhne here treated of, has taken the
liberty to lay before the reader the account which this family have
written of themselves".

I can go and check Ó Muraíle's edition for clarification. Or we could
even look out the original ms. on microfilm, since it is in the RIA...