Phase 2 Complete / erroneous material moved
It is my opinion that the editors of LL VI have messed up here and included the following material in this pedigree in error: "Araide Bibre in cáinte de Mumnechaib ba sé ba rectaire do Chormac .h. Chuind. Cairech a ben. Is i ro anacht Fiacha mc Oengusa. inde dicitur Fiacha Araide a quo Dal Araide." This appears in a text box at the bottom of the page, with a call-mark replicated next to "Feradaig" (loegaire.ttl) in column e. The combined effect is to produce a distant descendant of Níall Noígiallach who was contemporary with Cormac mac Airt. This is, of course, impossible. 

However, the call-mark is exactly level with Fiacha Araide (dail_araide.ttl) in column g and, as the note is basically a gloss on this name, I feel that it is to that that it refers. I am therefore going to transfer the relevant data across to dáil_araide.ttl, leaving a continuous and logical pedigree here. Order is restored.