Phase 2 Complete
This tract compares very oddly with the other tract of the same name (corco_bascind.ttl). There is a lot of overlap but corco_bascind.ttl contains multiple double-barreled names where this tract contains two individuals. This is all going to look a bit messy on the graph, as presumably both individuals in this tract are owl:sameAs the single individual in the other tract?