LMisc 610 ongoing
Section 14 complete.

This section is very terse and the connections between characters are not always apparent. It should be reviewed against the IrishGen database and other sources once more material has been added. Also, a translation exists of 'Conailla medb michura' in Études Celtiques 29; this will be consulted when the opportunity permits. EPT